test: lint fixtures
