+++ test/fixtures/test-runner/coverage-snap/b.js
@@ -1,0 +1,11 @@
+'use strict';
+// Here we can't import common module as the coverage will be different based on the system
+
+// Empty functions that don't do anything
+function doNothing1() {
+  // Not implemented
+}
+
+function doNothing2() {
+  // No logic here
+}
