--- test/fixtures/test-runner/coverage-snap/b.mjs
@@ -1,8 +1,0 @@
-// Empty functions that don't do anything
-function doNothing1() {
-    // Not implemented
-}
-
-function doNothing2() {
-    // No logic here
-}
