+++ test/fixtures/test-runner/output/coverage-width-100-uncovered-lines.snapshot
@@ -15,13 +15,13 @@ ok 1 - Coverage Print Fixed Width 100
 # duration_ms *
 # start of coverage report
 # --------------------------------------------------------------------------------------------------
-# file                        | line % | branch % | funcs % | uncovered lines
+# file     | line % | branch % | funcs % | uncovered lines
 # --------------------------------------------------------------------------------------------------
-# …runner/coverage-snap/a.mjs |  53.06 |   100.00 |    0.00 | 2-4 6-8 10-12 14-16 26-27 37-39 42-4…
-# …runner/coverage-snap/b.mjs |  25.00 |   100.00 |    0.00 | 2-4 6-8
-# …p/many-uncovered-lines.mjs |  59.11 |    42.86 |    3.23 | 2-4 6-8 10-12 14-16 26-27 37-39 42-4…
-# …dth-100-uncovered-lines.js | 100.00 |   100.00 |  100.00 | 
+# …ap/a.js |  55.77 |   100.00 |    0.00 | 5-7 9-11 13-15 17-19 29-30 40-42 45-47 50-52
+# …ap/b.js |  45.45 |   100.00 |    0.00 | 5-7 9-11
+# …ines.js |  50.99 |    42.86 |    1.92 | 5-7 9-11 13-15 17-19 29-30 40-42 45-47 50-52 55-57 59-6…
+# …nes.mjs | 100.00 |   100.00 |  100.00 | 
 # --------------------------------------------------------------------------------------------------
-# all files                   |  58.42 |    60.00 |    2.44 |
+# all fil… |  52.80 |    60.00 |    1.61 |
 # --------------------------------------------------------------------------------------------------
 # end of coverage report
