+++ test/fixtures/test-runner/output/coverage-width-100.snapshot
@@ -17,10 +17,10 @@ ok 1 - Coverage Print Fixed Width 100
 # --------------------------------------------------------------------------------------------------
 # file                                             | line % | branch % | funcs % | uncovered lines
 # --------------------------------------------------------------------------------------------------
-# test/fixtures/test-runner/coverage-snap/a.mjs    |  53.06 |   100.00 |    0.00 | 2-4 6-8 10-12 1…
-# test/fixtures/test-runner/coverage-snap/b.mjs    |  25.00 |   100.00 |    0.00 | 2-4 6-8
-# …xtures/test-runner/output/coverage-width-100.js | 100.00 |   100.00 |  100.00 | 
+# test/fixtures/test-runner/coverage-snap/a.js     |  55.77 |   100.00 |    0.00 | 5-7 9-11 13-15 …
+# test/fixtures/test-runner/coverage-snap/b.js     |  45.45 |   100.00 |    0.00 | 5-7 9-11
+# …tures/test-runner/output/coverage-width-100.mjs | 100.00 |   100.00 |  100.00 | 
 # --------------------------------------------------------------------------------------------------
-# all files                                        |  55.38 |   100.00 |    0.00 |
+# all files                                        |  60.81 |   100.00 |    0.00 |
 # --------------------------------------------------------------------------------------------------
 # end of coverage report
