--- test/fixtures/test-runner/output/coverage-width-80-uncovered-lines.js
@@ -1,9 +1,0 @@
-// Flags: --experimental-test-coverage
-import * as a from '../coverage-snap/b.mjs';
-import * as b from '../coverage-snap/a.mjs';
-import * as c from '../coverage-snap/many-uncovered-lines.mjs';
-import { test } from 'node:test';
-
-process.stdout.columns = 80;
-
-test(`Coverage Print Fixed Width ${process.stdout.columns}`);
