+++ test/fixtures/test-runner/output/coverage-width-80.snapshot
@@ -15,12 +15,12 @@ ok 1 - Coverage Print Fixed Width 80
 # duration_ms *
 # start of coverage report
 # ------------------------------------------------------------------------------
-# file                                        | line % | branch % | funcs % | …
+# file                                         | line % | branch % | funcs % | …
 # ------------------------------------------------------------------------------
-# …t/fixtures/test-runner/coverage-snap/a.mjs |  53.06 |   100.00 |    0.00 | …
-# …t/fixtures/test-runner/coverage-snap/b.mjs |  25.00 |   100.00 |    0.00 | …
-# …es/test-runner/output/coverage-width-80.js | 100.00 |   100.00 |  100.00 | 
+# test/fixtures/test-runner/coverage-snap/a.js |  55.77 |   100.00 |    0.00 | …
+# test/fixtures/test-runner/coverage-snap/b.js |  45.45 |   100.00 |    0.00 | …
+# …es/test-runner/output/coverage-width-80.mjs | 100.00 |   100.00 |  100.00 | 
 # ------------------------------------------------------------------------------
-# all files                                   |  55.38 |   100.00 |    0.00 |
+# all files                                    |  60.81 |   100.00 |    0.00 |
 # ------------------------------------------------------------------------------
 # end of coverage report
