+++ test/fixtures/test-runner/output/coverage-width-infinity-uncovered-lines.snapshot
@@ -14,14 +14,14 @@ ok 1 - Coverage Print Fixed Width Infinity
 # todo 0
 # duration_ms *
 # start of coverage report
-# ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-# file                                                                        | line % | branch % | funcs % | uncovered lines
-# ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-# test/fixtures/test-runner/coverage-snap/a.mjs                               |  53.06 |   100.00 |    0.00 | 2-4 6-8 10-12 14-16 26-27 37-39 42-44 47-49
-# test/fixtures/test-runner/coverage-snap/b.mjs                               |  25.00 |   100.00 |    0.00 | 2-4 6-8
-# test/fixtures/test-runner/coverage-snap/many-uncovered-lines.mjs            |  59.11 |    42.86 |    3.23 | 2-4 6-8 10-12 14-16 26-27 37-39 42-44 47-49 52-54 56-58 60-62 64-66 88-90 93-95 97-99 101-103 108-109 115-116 119-120 124-126 129-133 141-143 147 163-164 170 177 185-186 193-197 204-210 213-215 218-220 223-225
-# test/fixtures/test-runner/output/coverage-width-infinity-uncovered-lines.js | 100.00 |   100.00 |  100.00 | 
-# ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-# all files                                                                   |  58.42 |    60.00 |    2.44 |
-# ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+# file                                                                         | line % | branch % | funcs % | uncovered lines
+# ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+# test/fixtures/test-runner/coverage-snap/a.js                                 |  55.77 |   100.00 |    0.00 | 5-7 9-11 13-15 17-19 29-30 40-42 45-47 50-52
+# test/fixtures/test-runner/coverage-snap/b.js                                 |  45.45 |   100.00 |    0.00 | 5-7 9-11
+# test/fixtures/test-runner/coverage-snap/many-uncovered-lines.js              |  50.99 |    42.86 |    1.92 | 5-7 9-11 13-15 17-19 29-30 40-42 45-47 50-52 55-57 59-61 63-65 67-69 91-93 96-98 100-102 104-106 111-112 118-119 122-123 127-129 132-136 144-146 150 166-167 173 180 188-189 196-200 207-213 216-218 221-223 226-228 232 236-238 241-243 246-248 251-257 260-262 265-268 271-273 276-280 283-285 288-290 293-295 298-301 304-306 309-312 315-318 321-324 327-329 332-340 343-348 351-353
+# test/fixtures/test-runner/output/coverage-width-infinity-uncovered-lines.mjs | 100.00 |   100.00 |  100.00 | 
+# ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+# all files                                                                    |  52.80 |    60.00 |    1.61 |
+# ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 # end of coverage report
