--- test/fixtures/test-runner/output/coverage-width-infinity.js
@@ -1,8 +1,0 @@
-// Flags: --experimental-test-coverage
-import * as a from '../coverage-snap/b.mjs';
-import * as b from '../coverage-snap/a.mjs';
-import { test } from 'node:test';
-
-process.stdout.columns = Infinity;
-
-test(`Coverage Print Fixed Width ${process.stdout.columns}`);
