+++ test/fixtures/test-runner/output/coverage-width-infinity.snapshot
@@ -14,13 +14,13 @@ ok 1 - Coverage Print Fixed Width Infinity
 # todo 0
 # duration_ms *
 # start of coverage report
-# ----------------------------------------------------------------------------------------------------------------------------------------
-# file                                                        | line % | branch % | funcs % | uncovered lines
-# ----------------------------------------------------------------------------------------------------------------------------------------
-# test/fixtures/test-runner/coverage-snap/a.mjs               |  53.06 |   100.00 |    0.00 | 2-4 6-8 10-12 14-16 26-27 37-39 42-44 47-49
-# test/fixtures/test-runner/coverage-snap/b.mjs               |  25.00 |   100.00 |    0.00 | 2-4 6-8
-# test/fixtures/test-runner/output/coverage-width-infinity.js | 100.00 |   100.00 |  100.00 | 
-# ----------------------------------------------------------------------------------------------------------------------------------------
-# all files                                                   |  55.38 |   100.00 |    0.00 |
-# ----------------------------------------------------------------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------------------------------------------------------------------
+# file                                                         | line % | branch % | funcs % | uncovered lines
+# ------------------------------------------------------------------------------------------------------------------------------------------
+# test/fixtures/test-runner/coverage-snap/a.js                 |  55.77 |   100.00 |    0.00 | 5-7 9-11 13-15 17-19 29-30 40-42 45-47 50-52
+# test/fixtures/test-runner/coverage-snap/b.js                 |  45.45 |   100.00 |    0.00 | 5-7 9-11
+# test/fixtures/test-runner/output/coverage-width-infinity.mjs | 100.00 |   100.00 |  100.00 | 
+# ------------------------------------------------------------------------------------------------------------------------------------------
+# all files                                                    |  60.81 |   100.00 |    0.00 |
+# ------------------------------------------------------------------------------------------------------------------------------------------
 # end of coverage report
